moved all tests to test directory
